Changed subtitle to fit content
Changed "Find the sum of a set of consecutive numbers. This bot will calculate the sum of the consecutive numbers from 1 to any number you input below:"  to "Find the sum of a set of consecutive numbers. This bot will calculate the sum of consecutive numbers. Just type in the starting and ending numbers below:".